align the positioning version used by the poi poc to the one used by navigation

+++ src/poi-service/poi-common/common-data-model.h
@@ -1093,9 +1093,9 @@ public:
     {
         double latitude;
         double longitude;
-        int altitude;
+        double altitude;
     };
-    typedef ::DBus::Struct< double, double, int32_t > DBus_geoCoordinate3D_t;
+    typedef ::DBus::Struct< double, double, double > DBus_geoCoordinate3D_t;
 
     DBus_geoCoordinate3D(){
         m_geoCoordinate3D.latitude = 48.85792; //by default center of Paris
